@@ -12,54 +12,55 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.4.6",
-    "@angular/common": "^4.2.4",
-    "@angular/compiler": "^4.2.4",
-    "@angular/core": "^4.2.4",
-    "@angular/forms": "^4.2.4",
-    "@angular/http": "^4.2.4",
-    "@angular/platform-browser": "^4.2.4",
-    "@angular/platform-browser-dynamic": "^4.2.4",
-    "@angular/router": "^4.2.4",
-    "@ngx-translate/core": "^8.0.0",
-    "@ngx-translate/http-loader": "^2.0.0",
-    "angular-sortablejs": "^2.2.1",
-    "angular2-websocket": "^0.9.5",
-    "bootstrap": "^4.0.0-beta.2",
-    "bootstrap-slider": "^10.0.2",
-    "core-js": "^2.4.1",
-    "file-saver": "^1.3.3",
+    "@angular/animations": "^6.0.9",
+    "@angular/common": "^6.0.9",
+    "@angular/compiler": "^6.0.9",
+    "@angular/core": "^6.0.9",
+    "@angular/forms": "^6.0.9",
+    "@angular/http": "^6.0.9",
+    "@angular/platform-browser": "^6.0.9",
+    "@angular/platform-browser-dynamic": "^6.0.9",
+    "@angular/router": "^6.0.9",
+    "@ngx-translate/core": "^10.0.2",
+    "@ngx-translate/http-loader": "^3.0.1",
+    "angular-sortablejs": "^2.5.2",
+    "angular2-websocket": "^0.9.6",
+    "bootstrap": "^4.1.2",
+    "core-js": "^2.5.7",
+    "file-saver": "^1.3.8",
     "font-awesome": "^4.7.0",
-    "jquery": "^3.2.1",
-    "ngx-bootstrap": "^2.0.0-beta.11",
+    "jquery": "^3.3.1",
+    "ngx-bootstrap": "^2.0.5",
     "ngx-dropzone-wrapper": "^4.9.0",
-    "ngx-toastr": "^6.4.1-beta.0",
-    "rxjs": "^5.4.2",
+    "ngx-toastr": "^8.9.1",
+    "rxjs": "^6.2.2",
+    "rxjs-compat": "^6.0.0-rc.0",
     "sortablejs": "^1.7.0",
-    "tether": "^1.4.0",
+    "tether": "^1.4.4",
     "web-animations-js": "^2.3.1",
-    "zone.js": "^0.8.14"
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "1.3.2",
-    "@angular/compiler-cli": "^4.2.4",
-    "@angular/language-service": "^4.2.4",
+    "@angular-devkit/build-angular": "~0.6.8",
+    "@angular/cli": "^6.0.8",
+    "@angular/compiler-cli": "^6.0.9",
+    "@angular/language-service": "^6.0.9",
     "@types/jasmine": "~2.5.53",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~3.1.1",
+    "@types/jasminewd2": "^2.0.3",
+    "@types/node": "^6.0.114",
+    "codelyzer": "4.3.0",
     "jasmine-core": "~2.6.2",
     "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
+    "karma": "^1.7.1",
     "karma-chrome-launcher": "~2.1.1",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "karma-coverage-istanbul-reporter": "^1.4.3",
+    "karma-jasmine": "^1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
-    "node-sass": "^4.6.1",
+    "node-sass": "^4.9.2",
     "protractor": "~5.1.2",
     "ts-node": "~3.2.0",
     "tslint": "~5.3.2",
-    "typescript": "~2.3.3"
+    "typescript": "~2.7.2"
   }
 }
